requirements.txt
Pillow>=9.0.0
opencv-python>=4.8.0
tqdm>=4.64.0
psutil>=5.9.0
python-dotenv>=1.0.1
